Refactor properties::config into separate class
datatools::properties couples its C++ representation and format
specification through the inner datatools::properties::config class.
This makes it difficult to extend or modify the reading/writing (e.g. to
support databases) functionality, and to decouple aspects of parsing
such as file/variant lookup. The parsing is also the most complex part
of the properties code, whilst the C++ representation of a properties
object is very straightforward.

Split the datatools::properties::config class out into a dedicated
datatools::properties_config class. Update interface of
datatools::properties::data to provide the <type>_label functions rather
than as hidden _format methods. This is a grey area between type/text
representation. Making these functions part of data provides the
simplest decoupling for now. Update implementation of reading/writing
implementations to no longer require friendship between properties and
properties_config.

Update clients of datatools::properties::config to use
datatools::properties_config.
